Time: 34 ms (49.64%) | Memory: 28.8 MB (70.04%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cs
@@ -25,7 +25,7 @@ public class Solution {
     public bool JudgeSquareSum(int c) {
         
         int left = 0;
-        int right =  (int)Math.Sqrt(c);
+        long right =  (int)Math.Sqrt(c);
         
         while (left<=right){
